Clarify Underlying Issue When Missing RESTART DATE in SKIPREST
When a restarted model uses SKIPREST, there must be DATES or TSTEP
records/report times that correspond exactly to the restart time.
The current diagnostic message is a little too brief and developer
centric and does not offer sufficient hints to the user as to what
the underlying issue might be.

Try to expand the current message and provide more clues to the user
by mentioning both the SKIPREST and RESTART keywords in addition to
the expected and encountered time points.

Example original message:
  Error: Problem with keyword DATES
  In PRED_FLOW.DATA line 223
  At date: 2019-04-18 - scanned past restart data: 2019-04-12

Example message in this commit:
  Error: Problem with keyword DATES
  In PRED_FLOW.DATA line 223
  In a restarted simulation using SKIPREST, the DATES keyword must have
  a record corresponding to the RESTART time 12-Apr-2019 00:00:00.
  Reached time 18-Apr-2019 00:00:00 without an intervening record.