Disallow Wells Parented Directly to FIELD
The commercial simulator does not support parenting a well directly
to the FIELD group and Flow in multiple locations implicitly assumes
that no well is parented directly to FIELD either.

With this commit we detect this situation at the parser stage and
stop reading the input file.  In this case, we issue a diagnostic
message of the form

    Error: Problem with keyword WELSPECS
    In SPE1CASE1.DATA line 384
    Wells cannot be parented directly to 'FIELD'.
    Wells entered with disallowed 'FIELD' parent group:
     * PROD
     * INJ

    Error: Unrecoverable errors were encountered while loading input