Pull Basic Connection Information Out To Helper
This commit introduces a new helper class, WellConnectionRecord,
which holds the connection IJK and host grid information.  Said
information must be output if the run requests either RFT, PLT, or
both RFT and PLT data and so should not be tied to the RFT writer.

This is in preparation of adding PLT output.