Add Basic PLT Information to RFT Output File
We support both standard and multisegmented wells, but do not output
every known PLT array.  Notably missing are

  - WELLPLT
  - Brine flow rate (CONSRAT)
  - Phase flow rates through tubing (CONxTUB)

Computing CONNXT for multisegmented wells is somewhat involved, and
requires sorting the connections according to branch number,
segments, and tubing lengths.  We use an ancillary structure, an
array of indices into a WellConnections object, along with a custom
predicate to perform this sorting.

As an implementation detail, we handle PLT data for multisegmented
wells as an extension of the PLT data for regular/standard wells.
This extension is realised through dynamic polymorphism.

Add unit tests to cover PLT data for both well types.