Add Facility For Identifying Specific Well Segments
The primary use case is identifying UDQ sets when the model defines
UDQs at the segment level (SU*).  The call site will construct a
SegmentMatcher object backed by a specific ScheduleState, and then
segment-level UDQs will match against this to determine the correct
UDQ set on which to define the quantity.