Enumerate Columns According to Horizontal Model Dimensions
The inner loop is always across the model layers while we ensure
that the 'outer' loop always iterates over an index range that is at
least as large as that of the 'middle' loop.  This follows from
recent work on compatibility prompted by real field models with the
constant flux analytic aquifer type.

Thanks to [at]tskille for the additional insight.