Restore Builds With Dune in Non-Default Installed Locations
In general we'd need to link to dune-common ("target_link_library")
to get the appropriate include directories for Dune in that case,
but in this specific instance we can get by with not specialising
Dune's 'className' function if Dune is not available.

Thanks to [at]akva2 for suggesting this workaround.