Dimension Segment Related Restart Arrays According to Dynamic Size
The [IR]SEG and ILB[RS] arrays must be able to accommodate the
maximum number of segments and branches used in the run.  This
commit incorporates the dynamic maximum sizes.  If those sizes
exceed the maximum values entered in WSEGDIMS, then the resulting
restart file will not be fully compatible with other simulation
software.