Add Facility for Recording ACTIONX Well Structure Changes
This commit adds a new member, well_structure_change, to the
SimulatorUpdate structure.  The member defaults to 'false', but
will be set to 'true' if an ACTIONX block contains at least one of
a select group of keywords that affect the model's well topology.

In particular, set this member to 'true' if the ACTIONX block has
at least one of the keywords

  - COMPDAT
  - WELOPEN
  - WELSPECS

This will enable adding simulator logic to open or create wells in
the middle of a report step.