Add Emplace-like Operation to OrderedMap
This isn't quite the same as the general 'emplace()' member function
of std::map/std::unordered_map, because OrderedMap::emplace() knows
that the key type is string-like and therefore uses a distinct
parameter for the key.  We also take special care to always update
the 'key' part even when not inserting a new element.  The reason is
that the 'emplace()' call might have used a different yet equivalent
key to one added earlier.  In this case, the new key should override
the existing key.

This is in preparation of making FIP* array names unique only up to
the first three characters after the "FIP" prefix.