Prevent Duplicate Inlet Segments
If one WELSEGS keyword does not redefine every existing segment,
then we risk adding the same inlet segment multiple times.  If that
happens, we get a segment structure for which there might appear to
be multiple inlet segments on the same branch which is not possible
in a tree structure.  This commit prevents that situation by only
adding the new segment number if it is not already listed in
'm_inlet_segments'.