Separate New vs Update WELSPECS Handling out to Helpers
This commit creates new helper functions

    Schedule::welspecsCreateNewWell()
    Schedule::welspecsUpdateExistingWells()

which handle the creation of a new well or the property update of
one or more existing wells, respectively, in the context of
processing records in the WELSPECS keyword.

In preparation of making the WELSPECS keyword useful for changing
the controlling group of a collection of wells, we pass a vector of
well names as an argument to the update function.  Furthermore, we
construct this vector using normal 'wellNames()' processing which
enables using patterns and well list names such as

  * 'P-2'
  * 'PROD-*'
  * '?'
  * '*PROD'

as arguments/well names (item 1).