Support SWATINIT-Like Rescaling of PCOW at Restart
This commit adds a new, very specialised, operation to the material
law manager,

  void EclMaterialLawManager::applyRestartSwatInit(cell, maxPcow)

This will apply a SWATINIT-like rescaling of the oil/water capillary
pressure curve based on a caller-provided maximum capillary pressure
value.  The primary use case is this maximum value being taken from
the PPCW array in a restart file at simulation restart time.  We
assign the "maxPcow" member of the associate EpsInfo structure and
reinitialise the EpsPoints structure using this new information.
The latter is needed lest the maxPcnw() member function return
incorrect values in restarted simulations.