Extract CTF Input Data to Separate Structure
This commit creates a new helper structure,

    Opm::Connection::CTFProperties

that holds all of the individual items which go into calculating the
final connection transmissibility factor (CTF).  These include the
'Kh' product, the pressure equivalent radius ('r0'), the well-bore
radius ('rw'), and the skin factor.  We reimplement the Connection
constructor in terms of this new helper structure and this changes
the API of the constructor.  As an additional helper for the
implementation, and with a view towards simulation restart support,
we also store the value of the Peaceman formula denominator in
CTFProperties as

    double CTFProperties::peaceman_denom

This, in turn, simplifies the implementation of CSKIN.  It also
fixes a latent problem in the current CSKIN implementation which
would ignore any explicitly input CTF values in favour of the
pressure equivalent radius.  By updating the explicit denominator
with a new skin factor instead we do not need to recompute the
logarithmic term involving 'r0' as part of CSKIN handling.