Move Static Dake Model Calculation to Connection Class
Certain parts of the Dake model D-factor correlation relation depend
only on static parameters at the cell level.  Calculate the static
product once, and store it as a member of Connection::CTFProperties.
In particular, the new member

    double CTFProperties::static_dfac_corr_coeff

stores the product

    A * (K/K0)**B * poro**C * Ke/(h*rw)

with 'A', 'B', and 'C', being the Dake model parameters.  We compute
new values for the 'static_dfac_corr_coeff' as part of processing
the WDFACCOR keyword.

This static product will also be output to the restart file in
follow-up commits, and will be used to simplify evaluating the full
dynamic Dake model correlation.